Fix problèmes de compilation

+++ BSA IA/Algorithm.h
@@ -97,7 +97,7 @@ public:
 
 	const vector<Solution*>& current_solutions() const;
 
-	unsigned double global_best_cost() const;
+	double global_best_cost() const;
 
 	Solution& solution(const unsigned int index) const;
